updated contracts with new entrypoint

+++ account-abstraction-contracts/contracts/Implementation/SimpleAccountFactory.sol
@@ -25,12 +25,12 @@ contract SimpleAccountFactory {
      * Note that during UserOperation execution, this method is called only if the account is not deployed.
      * This method returns an existing account address so that entryPoint.getSenderAddress() would work even after account creation
      */
-    function createAccount(address owner, uint256 salt)
-        public
-        returns (SimpleAccount ret)
-    {
+    function createAccount(
+        address owner,
+        uint256 salt
+    ) public returns (SimpleAccount ret) {
         address addr = getAddress(owner, salt);
-        uint256 codeSize = addr.code.length;
+        uint codeSize = addr.code.length;
         if (codeSize > 0) {
             return SimpleAccount(payable(addr));
         }
@@ -47,11 +47,10 @@ contract SimpleAccountFactory {
     /**
      * calculate the counterfactual address of this account as it would be returned by createAccount()
      */
-    function getAddress(address owner, uint256 salt)
-        public
-        view
-        returns (address)
-    {
+    function getAddress(
+        address owner,
+        uint256 salt
+    ) public view returns (address) {
         return
             Create2.computeAddress(
                 bytes32(salt),
